+++ src/main/java/com/eventux/backend/model/Files.java
@@ -15,7 +15,7 @@ public class Files {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long fileID;
+    private Integer fileID;
 
     private String FileName;
